requirements.txt
Django
Fiona
geopy
pyproj
requests
Shapely
urllib3
Pillow
django-leaflet
django-nested-admin
django-crispy-forms
psycopg2-binary
openpyxl
djangorestframework
djangorestframework-gis
django-microsoft-auth
django-six
django-filter==2.0.0
